Factor detectable environments to platforms for #6.
Instead of hardcoding all the supported hosting environments
together into the main module, define an API allowing definition
and registration of hosting environments. This makes things
much cleaner as related code can all be kept together instead
of intermixed with the code of unrelated platforms. This also
enables an amount of dependency injection.