Add platform options passing mechanism.
Discussed in #6. Instead of using the PORT envvar and
calling app.get('port') to try to determine the port to
host the app on for the http module, now use the new options
platform property. This is done in such a way to support
the scenario where an app is being deployed into a hosting
environment where autoserve is rigged by an outside module.