Keep module-level and platform-level APIs separate.
In #6, it is discussed how the user can either call
detect() and manage the returned platform directly or
call the module directly as convenience. To support the
whole getBaseUrl(request) thing at the module level,
the module has to store a global indicating the platform
used (or maybe something like callback-local-storage
could be investigated). This change ensures that code
which chooses to manage the return value of detect()
directly can actually benefit from that by, e.g.,
hosting two different apps in two different platforms
within the same node process if things are rigged up
correctly.