Replace autoserve.getBaseUrl() with request.baseUrl.
This is cleaner and make supporting require('autoserve')(app)
less hacky by not requiring it to store a global of the detected
platform. However, to allow platforms to be defined nicely by
just exporting platform.getBaseUrl(req), this requires detect()
to return a wrapper around platform which handles setting
request.baseUrl.