Set options through module property.
Discussed in #6.

Now the way to set options is to modify the options
property of the module. This is to support deployment
scripts scenario/dependency injection.

Also, this change means that platforms do not need
to check for this.options to discover options.
Instead, they more clearly/conveniently are passed
options as the second argument to their serve()
function.